% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmkb-utils.R
\name{annotate_pmkb_entries_as_tsg_or_onc}
\alias{annotate_pmkb_entries_as_tsg_or_onc}
\title{annotate_pmkb_entries_as_tsg_or_onc}
\usage{
annotate_pmkb_entries_as_tsg_or_onc(pmkb)
}
\arguments{
\item{pmkb}{(data.table) the pmkb data.table}
}
\value{
data.table with added TSG/ONC annotation for each gene
}
\description{
Internal function to annotate genes in the PMKB table as either TSG or ONC

We used this function to generate pmkb_tsv and it is here for the sake of documentation.
}
