% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotated_bcf_to_oncotable}
\alias{annotated_bcf_to_oncotable}
\title{annotated_bcf_to_oncotable}
\usage{
annotated_bcf_to_oncotable(
  annotated_bcf,
  filter = "PASS",
  keepeff = c("trunc", "cnadel", "cnadup", "complexsv", "splice", "inframe_indel",
    "fusion", "missense", "promoter", "regulatory", "mir"),
  id = "",
  verbose = FALSE
)
}
\arguments{
\item{annotated_bcf}{(character) path to file}

\item{filter}{(character) default PASS}

\item{keepeff}{(character) variants to keep}

\item{id}{(character) ID of sample}

\item{verbose}{(logical) default FALSE}
}
\description{
Collect results from annotated BCF for oncotable
}
