% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file}
\alias{check_file}
\title{check_file}
\usage{
check_file(fn, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{fn}{(character) path}

\item{overwrite}{(logical) overwrite path even if existing, default FALSE}

\item{verbose}{(logical)}
}
\value{
TRUE if file exists and does not need to be regenerated, otherwise FALSE
}
\description{
check whether file needs to be generated/regenerated
}
