% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_GRanges_compatibility}
\alias{check_GRanges_compatibility}
\title{check_GRanges_compatibility}
\usage{
check_GRanges_compatibility(gr1, gr2, name1 = "first", name2 = "second")
}
\arguments{
\item{gr1}{GRanges object}

\item{gr2}{GRanges object}

\item{name1}{name of to describe gr1 (by default: 'first')}

\item{name2}{name of to describe gr2 (by default: 'second')}
}
\value{
TRUE if seqlevels are identical, otherwise, FALSE
}
\description{
Copmated the seqlevels of two GRanges objects. Returns TRUE/FALSE if the seqlevels are identical/non-identical and reports differences (in printed messages).
}
\author{
Alon Shaiber
}
