% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{cn.plot}
\alias{cn.plot}
\title{cn.plot}
\usage{
cn.plot(
  drivers.fname = NULL,
  complex.fname = NULL,
  cvgt.fname = "./coverage.gtrack.rds",
  gngt.fname = "./data/gt.ge.hg19.rds",
  cgcgt.fname = NULL,
  agt.fname = NULL,
  server = "",
  pair = "",
  ev.types = c("qrp", "tic", "qpdup", "qrdel", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm"),
  amp.thresh = 2,
  ploidy = 2,
  pad = 0.5,
  height = 1000,
  width = 1000,
  outdir = "./",
  verbose = TRUE
)
}
\arguments{
\item{drivers.fname}{text file with drivers + CN alterations + genomic locations}

\item{complex.fname}{(character) output from complex event caller}

\item{cvgt.fname}{(character) coverage gTrack file path}

\item{gngt.fname}{(character) gencode gTrack file path}

\item{cgcgt.fname}{(character) CGC gTrack file path}

\item{agt.fname}{(character) allele gTrack file name}

\item{server}{(character) path to gGnome.js server url}

\item{pair}{(character) pair id}

\item{ev.types}{(character) complex event types}

\item{amp.thresh}{(default 2) amplicon CN threshold for plotting ## possibly compute and save separately?}

\item{ploidy}{(default 2)}

\item{pad}{(numeric) window padding in bp default 5e5}

\item{height}{(numeric) plot height default 500}

\item{width}{(numeric) plot width default 500}

\item{outdir}{(character) where to save plots?}

\item{verbose}{(logical) default TRUE}
}
\value{
data.table with columns id, plot.fname, and plot.link
}
\description{
Creates .png files corresponding to gTrack plots for each amplified/deleted gene
}
