% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_rna_quantiles}
\alias{compute_rna_quantiles}
\title{compute_rna_quantiles}
\usage{
compute_rna_quantiles(
  tpm = NULL,
  tpm.cohort = NULL,
  onc = as.character(),
  tsg = as.character(),
  surface = as.character(),
  id = NULL,
  gencode.gtrack = NULL,
  quantile.thresh = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{tpm}{(character) path to file with RNA TPM input (can be kallisto, salmon, or prespecified format)}

\item{onc}{(character) character vector of oncogenes}

\item{tsg}{(character) character vector of TSGs}

\item{surface}{(character) character vector of surface genes}

\item{id}{(character) id of current sample}

\item{gencode.gtrack}{(character) path to gencode gTrack}

\item{quantile.thresh}{(numeric) default 0.05}

\item{verbose}{(logical)}

\item{tpm_cohort}{(character) path to file with cohort expression}
}
\value{
data.table with columns gene, pair, value, qt, role, and direction of over/under expression
}
\description{
Creates a data.table containing the expression level of each gene relative to a cohort.
If supplied, also adds role of gene (e.g. ONC/TSG)
}
