% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counts.to.allele.cn}
\alias{counts.to.allele.cn}
\title{counts.to.allele.cn}
\usage{
counts.to.allele.cn(counts, purity, ploidy)
}
\arguments{
\item{counts}{(numeric) numeric vector of counts}

\item{purity}{(numeric) purity}

\item{ploidy}{(numeric)}
}
\value{
cn (numeric) counts transformed to CN via purity + ploidy
}
\description{
allele rel2abs
}
\details{
zchoo Wednesday, May 05, 2021 11:22:17 AM
}
