% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covcbs}
\alias{covcbs}
\title{covcbs}
\usage{
covcbs(
  x,
  field = "ratio",
  name = "sample",
  max.ranges = 10000,
  lwd.border = 0.2,
  purity = NULL,
  ploidy = NULL,
  rebin = NULL,
  strip.chr = FALSE,
  ...
)
}
\arguments{
\item{x}{(character) path to coverage GRanges}

\item{field}{(character) field containing t/n ratio}

\item{name}{(character) name of gTrack}

\item{max.ranges}{(numeric) maximum ranges to plot, default 1e4}

\item{lwd.border}{(numeric) lwd of circles in gTrack, default 0.2}

\item{purity}{(numeric) sample purity estimate}

\item{ploidy}{(numeric) sample ploidy estimate}

\item{rebin}{(numeric) bin size (bp) if rebinning coverage}

\item{strip.chr}{(logical) strip chromosome prefix of coverage file}
}
\description{
covcbs
}
