% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create.summary}
\alias{create.summary}
\title{create.summary}
\usage{
create.summary(
  jabba_rds,
  snv_vcf,
  indel_vcf,
  verbose = FALSE,
  chrs = c(as.character(1:22), "X", "Y"),
  amp.thresh = 4,
  del.thresh = 0.5
)
}
\arguments{
\item{jabba_rds}{(character) jabba rds}

\item{snv_vcf}{(character) path to snv vcf/bcf}

\item{indel_vcf}{(character) path to indel vcf/bcf}

\item{verbose}{(logical) default FALSE}

\item{chrs}{(character vector) seqnames default 1:22, x, y}

\item{amp.thresh}{(numeric) default 4}

\item{del.thresh}{(numeric) default 0.5}
}
\value{
list with names
- purity
- ploidy
- junction_burden
- del_mbp
- amp_mbp
- cna_mbp
- cna_frac
- mut_count
- mut_per_mbp
}
\description{
calculate stuff for summary table
}
