% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{create_cgc_gtrack}
\alias{create_cgc_gtrack}
\title{create_cgc_gtrack}
\usage{
create_cgc_gtrack(cgc.fname = "./data/cgc.tsv", gencode.fname = NULL)
}
\arguments{
\item{cgc.fname}{(character)}

\item{gencode}{(character)}

\item{verbose}{(logical)}
}
\value{
file name of gencode gTrack
}
\description{
zchoo Monday, Apr 26, 2021 02:24:31 PM
}
