% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{create_genes_gtrack}
\alias{create_genes_gtrack}
\title{create_genes_gtrack}
\usage{
create_genes_gtrack(
  genes = c(),
  gencode.fname = NULL,
  coding.only = TRUE,
  level.thresh = 2,
  cached.dir = Sys.getenv("GENCODE_DIR")
)
}
\arguments{
\item{genes}{(character vector) genes to mark using "label" GRangesList metadata column (all other genes will have label == NA)}

\item{gencode.fname}{(character) path to GENCODE GFF}

\item{coding.only}{(logical) only include entries with gene_type == 'protein_coding'}

\item{level.thresh}{(numeric) only include entries with level <= level.thresh}

\item{cached.dir}{(character) path to directory in which to save the output gTrack}

\item{verbose}{(logical)}
}
\value{
file name of gencode gTrack
}
\description{
zchoo Monday, Apr 26, 2021 02:24:31 PM
}
