% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deconstructsigs_histogram}
\alias{deconstructsigs_histogram}
\title{deconstructsigs_histogram}
\usage{
deconstructsigs_histogram(
  sigs.fn = NULL,
  sigs.cohort.fn = NULL,
  id = "",
  cohort.type = "",
  outdir = "~",
  sigMet = NULL,
  ...
)
}
\arguments{
\item{sigs.fn}{(character) deconstructSigs results path}

\item{sigs.cohort.fn}{(character) path to cohort}

\item{cohort.type}{(character) e.g. supplied, Cell, tumor type ( actually optional )}

\item{sigMet}{(data.table) data table with description of mutaitonal signatures (sig.metadata.txt in the casereport data folder)}

\item{...}{additional params passed ppng}

\item{pair}{(character)}
}
\value{
histogram which you can then ppng etc.
}
\description{
TODO

create histogram plot for deconstructSigs
}
