% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmkb-utils.R
\name{extract_variant_types_from_pmkb}
\alias{extract_variant_types_from_pmkb}
\title{extract_variant_types_from_pmkb}
\usage{
extract_variant_types_from_pmkb(pmkb.csv = NA, output_file = NA)
}
\arguments{
\item{pmkb.csv}{(character) path to PMKB interpretations CSV (if not provided then the default table supplied with the package will be used). The table was downloaded from https://pmkb.weill.cornell.edu/about on 6/11/2021.}

\item{output_file}{(optional character) if supplied then the PMKB TSV will be saved to this path}
}
\value{
data.table
}
\description{
Internal function to parse the PMKB interpretations CSV and extract unique variant types

We used this function to generate pmkb_tsv and it is here for the sake of documentation.
pmkb_tsv is a data.table with the following columns: gene, variant, type, position.type, any.variant, gene.type.
gene - gene name (this column could contain duplicates since there will be a row for each variant value per gene
variant - a single variant description (the original PMKB interpretations CSV could have multiple variant description per line, separated by "|", we split these so that each variant appears in a separate line).
type - one of the following: missense, nonsense, frameshift, any mutation, insertion, deletion, copy number loss, stop gain, indel, copy number gain, rearrangement
position.type - either NA, codon, or exon. This is column is intended to identify these types of variants: "T618I" (type = codon), "codon(s) 515 missense" (type = codon), "exon(s) 34 frameshift" (type = exon)
any.variant - TRUE/FALSE. intended to identify variants such as: "any nonsense", "codon(s) 515 missense", "exon(s) 26-27 deletion", "any mutation"
gene.type - "TSG"/"ONC". We used a simple heuristic to annotate each gene as either ONC or TSG: see annotate_pmkb_entries_as_tsg_or_onc below for details.
}
