% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter.snpeff}
\alias{filter.snpeff}
\title{filter.snpeff}
\usage{
\method{filter}{snpeff}(
  vcf,
  gngt.fname,
  onc,
  tsg,
  drivers.fname,
  ref.name = "hg19",
  verbose = FALSE,
  type = NULL
)
}
\arguments{
\item{vcf}{(character) vcf file name}

\item{gngt.fname}{(character) gencode gTrack file name}

\item{onc}{(character) path to .rds file with character vector of oncogenes}

\item{tsg}{(character) path to .rds file with character vector of TSGs}

\item{ref.name}{(character) one of hg19 or hg38}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with columns gene, seqnames, pos, REF, ALT, variant.p, vartype, annotation
}
\description{
Overlaps annotated VCF with genes provided in delimited file and returns results as data.table
If VCF doesn't exist or is not provided returns an empty data.table
}
