% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.gallery.R
\name{fusion.plot}
\alias{fusion.plot}
\title{fusion.plot}
\usage{
fusion.plot(
  fs = NULL,
  complex.fname = NULL,
  cvgt.fname = NULL,
  agt.fname = NULL,
  gngt.fname = "/data/gt.ge.hg19.rds",
  pad = 1e+05,
  height = 1000,
  width = 1000,
  server = "",
  pair = "",
  outdir = "./"
)
}
\arguments{
\item{fs}{(gWalk) gWalk object containing filtered fusions}

\item{complex.fname}{(character)}

\item{cvgt.fname}{(character) coverage gTrack file name}

\item{agt.fname}{(character) allele gTrack file name}

\item{gngt.fname}{(character) gencode gTrack file name}

\item{pad}{(numeric) gWalk pad for plotting default 1e5}

\item{height}{(numeric) plot height default 1e3}

\item{width}{(numeric) plot width default 1e3}

\item{server}{(character) server url}

\item{pair}{(character) pair id}

\item{outdir}{(character) output directory}
}
\value{
gWalk with additional columns plot.fname (input to slickR)
}
\description{
Create .png files for each fusion
}
