% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.gallery.R
\name{fusion.table}
\alias{fusion.table}
\title{fusion.table}
\usage{
fusion.table(
  fusions.fname = NULL,
  complex.fname = NULL,
  cgc.fname = "/data/cgc.tsv",
  ev.types = c("qrp", "qpdup", "qrdel", "tic", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm")
)
}
\arguments{
\item{fusions.fname}{(character) file name containing fusion gWalks}

\item{complex.fname}{(character) file name to events}

\item{cgc.fname}{(character) cancer gene census file name}

\item{ev.types}{(character) event types}
}
\value{
gWalk (filtered) with metadata columns:
- walk.id
- name (e.g. involved genes)
- num.aa (total number of aminos)
- gene.pc (protein coordinates)
- driver (logical, is a gene a driver)
- ev.id (complex event IDs overlapping with walk)
- ev.type (complex event type overlapping with walk)
}
\description{
Prepare table of fusion events for displaying
}
