% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.gallery.R
\name{fusion.wrapper}
\alias{fusion.wrapper}
\title{fusion.wrapper}
\usage{
fusion.wrapper(
  fusions.fname = NULL,
  complex.fname = NULL,
  cvgt.fname = NULL,
  gngt.fname = NULL,
  agt.fname = NULL,
  cgc.fname = "/data/cgc.tsv",
  ev.types = c("qrp", "qpdup", "qrdel", "tic", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm"),
  height = 1000,
  width = 1000,
  server = "",
  pair = "",
  pad = 1e+05,
  outdir = "./"
)
}
\arguments{
\item{fusions.fname}{(character)}

\item{complex.fname}{(character)}

\item{cvgt.fname}{(character)}

\item{gngt.fname}{(character)}

\item{agt.fname}{(character)}

\item{cgc.fname}{(character)}

\item{ev.types}{(character)}

\item{height}{(numeric)}

\item{width}{(numeric)}

\item{server}{(character)}

\item{pair}{(character)}

\item{pad}{(numeric)}

\item{outdir}{(character)}
}
\value{
data.table
}
\description{
Wrapper for fusions
}
