% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{gallery.wrapper}
\alias{gallery.wrapper}
\title{gallery.wrapper}
\usage{
gallery.wrapper(
  complex.fname = NULL,
  background.fname = "/data/sv.burden.txt",
  cvgt.fname = "./coverage.gtrack.rds",
  gngt.fname = "./data/gt.ge.hg19.rds",
  cgcgt.fname = NULL,
  agt.fname = NULL,
  server = "",
  pair = "",
  ev.types = c("qrp", "tic", "qpdup", "qrdel", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm"),
  pad = 5e+05,
  height = 1000,
  width = 1000,
  outdir = "./"
)
}
\arguments{
\item{complex.fname}{(character)}

\item{background.fname}{(character) text file with event burdens (default sv.burden.txt)}

\item{cvgt.fname}{(character) coverage gTrack file path}

\item{gngt.fname}{(character) gencode gTrack file path}

\item{cgcgt.fname}{(character) CGC gTrack file path}

\item{agt.fname}{(character) allele gTrack file name}

\item{server}{(character) path to gGnome.js server url}

\item{pair}{(character) pair id}

\item{ev.types}{(character) complex event types}

\item{pad}{(numeric) window padding in bp default 5e5}

\item{height}{(numeric) plot height default 1000}

\item{width}{(numeric) plot width default 1000}

\item{outdir}{(character) where to save plots}
}
\value{
data.table with columns from complex and additionally plot.fname and plot.link
}
\description{
wrapper function to prep input for SlickR
}
