% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_ampdel_annotations}
\alias{get_gene_ampdel_annotations}
\title{zchoo Monday, May 03, 2021 02:42:51 PM}
\usage{
get_gene_ampdel_annotations(genes_cn, amp.thresh, del.thresh)
}
\description{
zchoo Monday, May 03, 2021 02:42:51 PM
}
