% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_copy_numbers}
\alias{get_gene_copy_numbers}
\title{get_gene_copy_numbers}
\usage{
get_gene_copy_numbers(
  gg,
  gene_ranges,
  complex.fname = NULL,
  nseg = NULL,
  gene_id_col = "gene_id",
  simplify_seqnames = FALSE,
  mfields = c("gene_name", "source", "gene_id", "gene_type", "level", "hgnc_id",
    "havana_gene"),
  ev.types = c("qrp", "qpdup", "qrdel", "tic", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm"),
  output_type = "data.table",
  ploidy = 2,
  verbose = TRUE
)
}
\arguments{
\item{gg}{either path to rds of a gGraph or an object containing the gGraph with JaBbA output}

\item{gene_ranges}{GRanges of genes (must contain field "gene_name"). Alternatively a path to a file that could be parsed by rtracklayer::import (such as gtf) is acceptable.}

\item{complex.fname}{(character) path to complex event calls}

\item{nseg}{GRanges with field "ncn" - the normal copy number (if not provided then ncn = 2 is used)}

\item{gene_id_col}{the name of the column to be used in order to identify genes (must be unique for each gene, so usually "gene_name" is not the right choice).}

\item{simplify_seqnames}{when set to TRUE, then gr.sub is ran on the seqnames of the gGraph segments and the genes GRanges}

\item{mfields}{the metadata fields that the output should inherit from the genes GRanges}

\item{ev.types}{complex event types (for now excluding simple dups/dels...)}

\item{output_type}{either GRanges or data.table}

\item{ploidy}{tumor ploidy default 2}
}
\value{
GRanges or data.table with genes CN
}
\description{
Takes a jabba_rds output and returns a GRanges with the genes that have either amplifications or deletions
}
\author{
Alon Shaiber
}
