% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_oncokb_annotations}
\alias{get_oncokb_annotations}
\title{get_oncokb_annotations}
\usage{
get_oncokb_annotations(
  oncokb.response,
  fields = c("highestSensitiveLevel", "highestResistanceLevel",
    "highestDiagnosticImplicationLevel", "highestPrognosticImplicationLevel")
)
}
\arguments{
\item{oncokb.response}{either a single response object or a list of response objects}

\item{fields}{the fields to extract from the OncoKB response}
}
\value{
data.table with the values of the fields in each response
}
\description{
Get a data.table containing specific fields from a list of OncoKB responses
}
\author{
Alon Shaiber
}
