% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_oncokb_gene_entry_url}
\alias{get_oncokb_gene_entry_url}
\title{get_oncokb_gene_entry_url}
\usage{
get_oncokb_gene_entry_url(oncokb.response)
}
\arguments{
\item{oncokb.response}{the JSON response, which is an output of the httr::GET query}
}
\value{
The URL if exists, otherwise NA
}
\description{
Get the url for the OncoKB entry for an actionable alteration
}
\author{
Alon Shaiber
}
