% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_oncokb_response}
\alias{get_oncokb_response}
\title{get_oncokb_response}
\usage{
get_oncokb_response(
  variants.dt,
  oncokb.token,
  oncokb.url = "https://www.oncokb.org/api/v1/annotate",
  reference = "GRCh37"
)
}
\arguments{
\item{variants.dt}{data.table with the variants to query. As a minimum, must contain the following columns: seqnames, start, end, REF, ALT.}

\item{oncokb.token}{a token to use when querying OncoKB. If you don't have a token, you must first obtain one from: https://www.oncokb.org/apiAccess}
}
\value{
list of response object. Response objects can be parsed using httr (for example: httr::content).
}
\description{
Using httr to query genomic changes on OncoKB in accordance with the web API specification in https://www.oncokb.org/swagger-ui/index.html .
}
\author{
Alon Shaiber
}
