% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmkb-utils.R
\name{get_pmkb_dt}
\alias{get_pmkb_dt}
\title{get_pmkb_dt}
\usage{
get_pmkb_dt(pmkb_tsv = NA)
}
\arguments{
\item{pmkb_tsv}{(character) path to PMKB TSV (if not provided then the default table supplied with the package will be used}
}
\description{
Internal function to read the pmkb TSV (this is the simplified version that we produced from the PMKB interpretations CSV)
}
