% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab.agtrack}
\alias{grab.agtrack}
\title{grab.agtrack}
\usage{
grab.agtrack(
  agt.fname = NULL,
  min.frac = 0.2,
  max.frac = 0.8,
  purity = NULL,
  ploidy = NULL,
  major.col = "red",
  minor.col = "blue",
  max.ranges = 10000,
  lwd.border = 0.2,
  name = "SNP",
  ...
)
}
\arguments{
\item{agt.fname}{(character) path to sites.txt}

\item{min.frac}{(numeric) between 0 and 1, min frequency in normal to count as het site}

\item{max.frac}{(numeric) between 0 and 1, max frequency in normal to count as het site}

\item{purity}{(numeric)}

\item{ploidy}{(numeric)}

\item{major.col}{(character) major allele color}

\item{minor.col}{(character) minor allele color}

\item{max.ranges}{(numeric) gTrack param}

\item{lwd.border}{(numeric)}

\item{name}{(character) gTrack name}

\item{...}{additional args to be passed to gTrack}
}
\value{
allele gTrack
}
\description{
returns allele gtrack given sites.txt from het pileup
}
