% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab.gene.ranges}
\alias{grab.gene.ranges}
\title{grab.gene.ranges}
\usage{
grab.gene.ranges(gngt.fname, genes = as.character())
}
\arguments{
\item{gngt.fname}{(character) gencode gTrack file name}

\item{genes}{(character) vector of gene symbols}
}
\value{
GRanges with metadata column gene_name
}
\description{
grab.gene.ranges
}
