% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab.hets}
\alias{grab.hets}
\title{grab.hets}
\usage{
grab.hets(agt.fname = NULL, min.frac = 0.2, max.frac = 0.8)
}
\arguments{
\item{agt.fname}{(character) path to sites.txt}

\item{min.frac}{(numeric) between 0 and 1, min frequency in normal to count as het site}

\item{max.frac}{(numeric) between 0 and 1, max frequency in normal to count as het site}
}
\value{
allele gTrack
}
\description{
returns allele gtrack given sites.txt from het pileup
}
