% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{grab.window}
\alias{grab.window}
\title{grab.window}
\usage{
grab.window(
  gr,
  complex.fname,
  amp.thresh = 4,
  ploidy = 2,
  ev.types = c(),
  return.type = "data.table"
)
}
\arguments{
\item{gr}{(GRanges)}

\item{complex.fname}{(character)}

\item{amp.thresh}{(numeric)}

\item{ploidy}{(numeric)}

\item{ev.types}{(character) list of acceptable event types (try to exclude simple events)}

\item{return.type}{(character) one of "data.table" or "GRanges"}
}
\value{
data.table with node.fp, amp.fp, and ev.fp for node, event, and amplicon footprints
}
\description{
helper function for adjusting the window around range of interest

if range falls within a complex event, then return the event footprint
if gene is amplified but not in a complex event but is in a complex amplicon:
- return amplicon
if gene is not in a complex event OR in a complex amplicon:
- return the range of the node(s) containing the gene
}
