% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSummaryTable}
\alias{makeSummaryTable}
\title{makeSummaryTable}
\usage{
makeSummaryTable(
  cnv_table,
  fusions_table,
  expression_table,
  mutations_table,
  onco_table
)
}
\arguments{
\item{cnv_table}{file path to casereport copy number variants table}

\item{fusions_table}{file path to casereport fusions table}

\item{expression_table}{file path to casereport over/under expression table}

\item{mutations_table}{file path to casereport driver mutations table}

\item{onco_table}{file path to casereport oncotable}

\item{the}{directory of casereport}
}
\value{
summary table of driver genes.
}
\description{
Function for generating a summary table of interesting driver genes for casereport.
}
