% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_expression_histograms}
\alias{plot_expression_histograms}
\title{plot_expression_histograms}
\usage{
plot_expression_histograms(
  rna.change.fn = NULL,
  melted.expr.fn = NULL,
  pair = NULL,
  outdir = ".",
  ...
)
}
\arguments{
\item{rna.change.fn}{(character) path to over/underexpressed oncogenes file}

\item{melted.expr.fn}{(character) path to melted expression file}

\item{pair}{(character)}

\item{outdir}{(character) where to save plots}

\item{...}{additional arguments to ppng}
}
\value{
data.table with columns $gene and $expr.hist.fname
}
\description{
Produce histograms for RNA expression TPM
}
