% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ppgrid}
\alias{ppgrid}
\title{ppgrid}
\usage{
ppgrid(
  segstats,
  allelic = FALSE,
  purity.min = 0.01,
  purity.max = 1,
  ploidy.step = 0.01,
  purity.step = 0.01,
  ploidy.min = 1.2,
  ploidy.max = 6,
  plot = F,
  verbose = F,
  mc.cores = 1,
  return.nll = FALSE
)
}
\arguments{
\item{segstats}{GRanges object of intervals with meta data fields "mean" and "sd" (i.e. output of segstats function)}

\item{allelic}{logical flag, if TRUE will also look for mean_high, sd_high, mean_low, sd_low variables and choose among top solutions from top copy number according to the best allelic fit}

\item{purity.min}{min purity value allowed}

\item{purity.max}{max purity value allowed}

\item{ploidy.step}{grid length of ploidy values}

\item{purity.step}{grid length of purity values}

\item{ploidy.min}{min ploidy value allowed}

\item{ploidy.max}{max ploidy value allowed}

\item{plot}{whether to plot the results to file}

\item{verbose}{print intermediate outputs}

\item{mc.cores}{integer number of cores to use (default 1)}

\item{return.all}{return the log likelihood matrix}
}
\value{
data.frame with top purity and ploidy solutions and associated gamma and beta values, for use in downstream jbaMI
}
\description{
least squares grid search for purity and ploidy modes
}
