% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rna.waterfall.plot}
\alias{rna.waterfall.plot}
\title{rna.waterfall.plot}
\usage{
rna.waterfall.plot(melted.expr.fn, rna.change.fn, pair)
}
\arguments{
\item{rna.change.fn}{(character) genes with chagne in RNA expression}

\item{pair}{character pair of interest. must be a column in tpm.cohort if tpm.pair is NULL}

\item{tpm.quantiles.fn}{file path data.table with column gene and tpm}
}
\value{
creates a plot
}
\description{
create waterfall plot
}
