% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rna_reformat}
\alias{rna_reformat}
\title{rna_reformat}
\usage{
rna_reformat(kallisto.fname, pair = NULL, gngt.fname = NULL)
}
\arguments{
\item{kallisto.fname}{(character) path to kallisto/salmon output}

\item{pair}{(character) id}

\item{gngt.fname}{(character) gencode gTrack file name}
}
\value{
data.table with columns gene, pair
}
\description{
preprocess kallisto/salmon raw tsv files for downstream analysis
}
