% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_param}
\alias{set_param}
\title{set_param}
\usage{
set_param(l = list(), name = "", value = NA_character_)
}
\arguments{
\item{l}{input list}

\item{name}{the name of the field}

\item{value}{the value to assign if the field is NULL (by default NA_character_)}
}
\value{
list
}
\description{
check if a field exists in a list and if not then set it
}
