% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.gallery.R
\name{star2grl}
\alias{star2grl}
\title{star2grl

Quick utility function to convert star-fusion breakpoints to GRangesList}
\usage{
star2grl(fname, chrsub = TRUE, return.type = "Junction", mc.cores = 16)
}
\arguments{
\item{fname}{(character) name of star-fusion output file}

\item{chrsub}{(logical) remove chr prefix default TRUE}

\item{return.type}{(character) Junction or GRangesList (default Junction)}

\item{mc.cores}{(numeric) default 16}

\item{GRangesList}{or Junction}
}
\description{
zchoo Tuesday, Apr 27, 2021 10:49:13 AM
this is to generate data tables and plots for fusions
}
