% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_cases}
\alias{summarize_cases}
\title{summarize_cases}
\usage{
summarize_cases(
  jb,
  output_file = NULL,
  libdir = "~/git/casereport",
  html_dir = NULL,
  metadata = NULL
)
}
\arguments{
\item{jb}{Flow Job object or a data.frame/data.table containing the contents of outputs(jb), where jb is the case report Flow Job object.}

\item{output_file}{output TXT in which the tabular data will be saved. If not file is provided then the data is saved to a temporary file.}

\item{libdir}{path to the casereport repository clone}

\item{html_dir}{path to the directory in which to put the html version of the table}

\item{metadata}{data.frame or data.table with additional metadata for the samples in jb. The metadata table must contain a header. The first column of the metadata should contain sample names that match the sample names in the Job object and should only contain unique values, otherwise this table is ignored and no metadata is added.}
}
\value{
data.table
}
\description{
Very ad-hoc function for mskilab use only!

takes a case report flow module and produces a data.table with links to the reports
this is assuming that the reports are somewhere under /gpfs/commons/projects/imielinski_web
}
