% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.gallery.R
\name{sv.plot}
\alias{sv.plot}
\title{sv.plot}
\usage{
sv.plot(
  complex.fname = NULL,
  cvgt.fname = "./coverage.gtrack.rds",
  gngt.fname = "./data/gt.ge.hg19.rds",
  genes.gt.fname = NULL,
  agt.fname = NULL,
  server = "",
  pair = "",
  ev.types = c("qrp", "tic", "qpdup", "qrdel", "bfb", "dm", "chromoplexy",
    "chromothripsis", "tyfonas", "rigma", "pyrgo", "cpxdm"),
  pad = 5e+05,
  height = 1000,
  width = 1000,
  outdir = "./"
)
}
\arguments{
\item{complex.fname}{(character) output from complex event caller}

\item{cvgt.fname}{(character) coverage gTrack file path}

\item{gngt.fname}{(character) gencode gTrack file path}

\item{genes.gt.fname}{(character) CGC gTrack file path}

\item{agt.fname}{(character) allele gTrack file name}

\item{server}{(character) server url}

\item{pair}{(character) pair id}

\item{ev.types}{(character) complex event types}

\item{pad}{(numeric) window padding in bp default 5e5}

\item{height}{(numeric) plot height default 500}

\item{width}{(numeric) plot width default 500}

\item{outdir}{(character) where to save plots?}
}
\value{
data.table with columns id, plot.fname, and plot.link
}
\description{
Creates .png files corresponding to gTrack plots for each complex SV pattern
}
