% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.gallery.R
\name{wgs.circos}
\alias{wgs.circos}
\title{wgs.circos

Quick utility function for circos plot with read depth, junctions, and segments
(copied from skitools)}
\usage{
wgs.circos(
  jabba_rds,
  cov_fn,
  transform = TRUE,
  segs = NULL,
  win = NULL,
  field = "ratio",
  cytoband = NULL,
  y.field = field,
  ylim = NA,
  cytoband.path = "~/DB/UCSC/hg19.cytoband.txt",
  cex.points = 1,
  ideogram.outer = TRUE,
  scatter = TRUE,
  bar = FALSE,
  line = FALSE,
  gap.after = 1,
  labels.cex = 1,
  y.quantile = 0.9999,
  chr.sub = TRUE,
  max.ranges = 10000,
  axis.frac = 0.02,
  palette = "BrBg",
  ...
)
}
\arguments{
\item{jabba_rds}{(character) path to jabba}

\item{cov_fn}{(character) path to coverage}

\item{transform}{(logical) rel2abs the coverage?}

\item{segs}{GRanges of segments with optional fields $col and $border}

\item{win}{GRanges window to limit plot to}

\item{cytoband}{GRanges of cytoband}

\item{y.field}{field in cov that specifies the y axis to draw}

\item{ylim}{ylim on cov (default automatically computed)}

\item{cytoband.path}{path to UCSC style cytoband path}

\item{cex.points}{cex for cov points}

\item{y.quantile}{quantile normalization}

\item{max.ranges}{max ranges for cov points (1e4)}

\item{chr.sum}{whether to chr.sub everything}
}
\description{
wgs.circos

Quick utility function for circos plot with read depth, junctions, and segments
(copied from skitools)
}
