% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casereport.R
\name{wgs.report}
\alias{wgs.report}
\title{wgs.report}
\usage{
wgs.report(opt)
}
\arguments{
\item{opt}{list with the input parameters for the wgs casereport. For more information on parameters run "Rscript wgs.report --help" in the command line.
add whether gene is TSG or ONCO}
}
\description{
Generate a WGS case report
}
