% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wgs_gtrack}
\alias{wgs_gtrack}
\title{wgs_gtrack}
\usage{
wgs_gtrack(jabba_rds, cvgt.fname, agt.fname = NULL)
}
\arguments{
\item{jabba_rds}{(character) path to jabba}

\item{cvgt.fname}{(character) coverage gTrack}

\item{agt.fname}{(character) allele gtrack}
}
\value{
gTrack object with nice formatting
}
\description{
wgs_gtrack
}
